fix printf string if size_t is bigger than int
On 64-bit systems, size_t might be larger than int so the compiler may
warn about printing a size_t using %d.  Since the length of result.data
will never be larger than an unsigned long, it is safe to downcast this
and use %lu instead.

+++ driver.c
@@ -32,7 +32,8 @@ int main(void)
 
 		rv = bc_decode(bits, &result, track);
 
-		printf("rv: %d, data_len: %d, data:\n`%s`\n", rv, strlen(result.data), result.data);
+		printf("rv: %d, data_len: %lu, data:\n`%s`\n",
+			rv, (unsigned long)strlen(result.data), result.data);
 
 		if (0 == rv)
 		{
